LayoutTests/inspector-obsolete/elements/elements-panel-search-expected.txt

Tests that elements panel search is returning proper results.

FooBar

Found by selector

Running: testSetUp

Running: testPlainText
F o o B a r

Running: testPartialText
F o o B a r

Running: testStartTag
< i n p u t   v a l u e = " I n p u t V a l " >

Running: testEndTag
< i n p u t   v a l u e = " I n p u t V a l " >

Running: testPartialTag
< i n p u t   v a l u e = " I n p u t V a l " >

Running: testPartialAbsentTagStart
Nothing found

Running: testPartialAbsentTagEnd
Nothing found

Running: testFullTag
< i n p u t   v a l u e = " I n p u t V a l " >

Running: testExactAttributeName
< i n p u t   v a l u e = " I n p u t V a l " >

Running: testExactAttributeVal_ue
< i n p u t   v a l u e = " I n p u t V a l " >

Running: testExactAttributeVal_ueWithQuotes
< d i v   a t t r = " f o o " > < / d i v >

Running: testPartialAttributeVal_ue
< i n p u t   v a l u e = " I n p u t V a l " >

Running: testXPathAttribute
< d i v   a t t r = " f o o " > < / d i v >

Running: testSelector
< s p a n > F o u n d   b y   s e l e c t o r < / s p a n >
